Minimal dataset to test BUSCO on human test data
BUSCO screens a genome for hits in a "lineage dataset", which have a
compressed size of many megabytes.  The human sequence data in this
repository (`homo_sapiens/genome/genome.fasta`) match a single entry of
the primate lineage dataset (`primates_odb10.2021-02-19.tar.gz`).

In the file submitted here, I removed all the non-matching entries of
the primate dataset, resulting in final size of 48 KiB.  I have verified
that it can be used to test BUSCO modules in less than 2 min.